+++ docs/releasenotes/releasenotes.rst
@@ -13,6 +13,11 @@ Enhancements
 
 * Google Maps: New block parameter ``showmarker`` which can be used to hide the marker by setting this parameter to ``false``
 
+Fixes
+~~~~~~
+
+* Setpoint devices: Fix for new Domoticz setpoint devices (Domoticz build 15555)
+
 V3.11.0 Beta (8-9-2023)
 ---------------------------